+++ src/main/java/com/plog/server/badge/api/MyBadgeController.java
@@ -23,11 +23,14 @@ public class MyBadgeController {
     @PatchMapping("/{userUUID}/{badgeId}")
     public ResponseEntity<ApiResponse<Void>> updateSelectedBadge(
             @PathVariable UUID userUUID,
-            @PathVariable Long badgeId) {
+            @PathVariable int badgeId) {
         myBadgeService.setSelectedBadge(userUUID, badgeId);
         // 응답 메시지 생성
         ApiResponse<Void> response = new ApiResponse<>("프로필 배지가 설정되었습니다.");
 
         return ResponseEntity.ok(response);
     }
+
+    //프로필 배지 조회
+
 }
