+++ src/main/java/com/plog/server/badge/service/MyBadgeService.java
@@ -19,13 +19,13 @@ public class MyBadgeService {
     private final ProfileRepository profileRepository;
     private final MyBadgeRepository myBadgeRepository;
     @Transactional
-    public void setSelectedBadge(UUID userUUID, Long badgeId) {
+    public void setSelectedBadge(UUID userUUID, int badgeId) {
         // 사용자 프로필 조회
         Profile profile = profileRepository.findByUserUserUUID(userUUID)
                 .orElseThrow(() -> new IllegalArgumentException("프로필이 없습니다: " + userUUID));
 
         // 해당 사용자가 소유한 배지인지 확인
-        MyBadge myBadge = myBadgeRepository.findByProfileAndBadgeBadgeId(profile, badgeId)
+        MyBadge myBadge = myBadgeRepository.findByProfileAndBadgeBadgeId(profile, (long) badgeId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 배지는 사용자가 소유하고 있지 않습니다."));
 
         // 프로필에 대표 배지 설정
@@ -34,4 +34,6 @@ public class MyBadgeService {
         // 프로필 저장
         profileRepository.save(profile);
     }
+
+
 }
